Extract resolvers and mutations for menu
